@@ -1,7 +1,7 @@
 import CloudKit
 
 /// Successful result of record modification and deletion functions, containing details about saved and deleted records.
-public struct ModifyRecordsResult: Equatable {
+public struct ModifyRecordsResult: Equatable, Sendable {
   /// An array of saved records. The records likely have different metadata from the records that you gave to the modification function
   /// as input, because CloudKit updates the record modification timestamp and change tag on the server side when saving records.
   public let savedRecords: [CKRecord]
