@@ -7,13 +7,14 @@ import Foundation
 /// You construct Canopy with injected CloudKit container and databases, token store, and settings provider.
 /// Canopy has reasonable defaults for all of these, and you need to only override the ones that need to use
 /// a different value from the default.
+@available(iOS 16.4, macOS 13.3, *)
 public actor Canopy: CanopyType {
-  private let containerProvider: () -> CKContainerType
-  private let publicCloudDatabaseProvider: () -> CKDatabaseType
-  private let privateCloudDatabaseProvider: () -> CKDatabaseType
-  private let sharedCloudDatabaseProvider: () -> CKDatabaseType
-  private let settingsProvider: () -> CanopySettingsType
-  private let tokenStoreProvider: () -> TokenStoreType
+  private let containerProvider: @Sendable () -> CKContainerType
+  private let publicCloudDatabaseProvider: @Sendable () -> CKDatabaseType
+  private let privateCloudDatabaseProvider: @Sendable () -> CKDatabaseType
+  private let sharedCloudDatabaseProvider: @Sendable () -> CKDatabaseType
+  private let settingsProvider: @Sendable () -> CanopySettingsType
+  private let tokenStoreProvider: @Sendable () -> TokenStoreType
   
   private var containerAPI: CKContainerAPI?
   private var databaseAPIs: [CKDatabase.Scope: CKDatabaseAPI] = [:]
@@ -31,12 +32,12 @@ public actor Canopy: CanopyType {
   ///   - tokenStore: an object that stores and returns zone and database tokens for the requests that work with the tokens.
   ///   Canopy only interacts with the token store when using the ``CKDatabaseAPIType/fetchDatabaseChanges(qualityOfService:)`` and ``CKDatabaseAPIType/fetchZoneChanges(recordZoneIDs:fetchMethod:qualityOfService:)`` APIs. If you don’t use these APIs, you can ignore this parameter.
   public init(
-    container: @escaping @autoclosure () -> CKContainerType = CKContainer.default(),
-    publicCloudDatabase: @escaping @autoclosure () -> CKDatabaseType = CKContainer.default().publicCloudDatabase,
-    privateCloudDatabase: @escaping @autoclosure () -> CKDatabaseType = CKContainer.default().privateCloudDatabase,
-    sharedCloudDatabase: @escaping @autoclosure () -> CKDatabaseType = CKContainer.default().sharedCloudDatabase,
-    settings: @escaping () -> CanopySettingsType = { CanopySettings() },
-    tokenStore: @escaping @autoclosure () -> TokenStoreType = UserDefaultsTokenStore()
+    container: @escaping @autoclosure @Sendable () -> CKContainerType = CKContainer.default(),
+    publicCloudDatabase: @escaping @autoclosure @Sendable () -> CKDatabaseType = CKContainer.default().publicCloudDatabase,
+    privateCloudDatabase: @escaping @autoclosure @Sendable () -> CKDatabaseType = CKContainer.default().privateCloudDatabase,
+    sharedCloudDatabase: @escaping @autoclosure @Sendable () -> CKDatabaseType = CKContainer.default().sharedCloudDatabase,
+    settings: @escaping @Sendable () -> CanopySettingsType = { CanopySettings() },
+    tokenStore: @escaping @autoclosure @Sendable () -> TokenStoreType = UserDefaultsTokenStore()
   ) {
     self.containerProvider = container
     self.publicCloudDatabaseProvider = publicCloudDatabase
